chore: rename imageUpload folder to upload
