+++ src/components/uploader/BaseImagePreview.vue
