+++ src/components/uploader/CloudinaryWidget.vue
